pyproject.toml
[tool.poetry]
name = "cortexquest"
version = "0.0.1"
description = "CortexQuest is a Python package designed to facilitate the reading, parsing, and analysis of questionnaires hosted on Google Sheets, specifically tailored for brain bank data. This tool aims to streamline the process of handling survey data related to neuroscience research, providing researchers with an efficient and user-friendly way to manage and analyze their data."
authors = ["Gal Kepler <galkepler@gmail.com>", "Tzlil Tabib <tzliltabib@gmail.com>"]
license = "MIT"
readme = "README.rst"
homepage = "https://github.com/GalKepler/CortexQuest"
repository = "https://github.com/GalKepler/CortexQuest"
documentation = "https://github.com/GalKepler/CortexQuest"
keywords = ["neuroscience", "brain bank", "google sheets", "survey", "analysis"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Information Analysis",
]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
typer = "^0.12.0"
tox = "^4.15.0"
pandas = "^2.2.2"
gspread = "^6.1.2"
oauth2client = "^4.1.3"
google-auth = "^2.29.0"
google-auth-oauthlib = "^1.2.0"
bump2version = "^1.0.1"
wheel = "^0.43.0"
sphinx = "^7.3.7"
twine = "^5.1.0"

translate = "^3.6.1"
ipykernel = "^6.29.4"
pre-commit = "^3.7.1"
chardet = "^5.2.0"
sphinx-rtd-theme = "^2.0.0"
numpy = "^1.26.4"
[tool.poetry.dev-dependencies]
coverage = "^7.5.1"  # testing
mypy = "^1.10.0"  # linting
pytest = "^8.2.1"  # testing
ruff = "^0.4.4"  # linting
black = "^24.4.2"
flake8 = "^7.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
cortexquest = "cortexquest:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.package-data]
"*" = ["*.*"]

# Mypy configuration
[tool.mypy]
files = "."

# Use strict defaults
strict = true
warn_unreachable = true
warn_no_return = true

[[tool.mypy.overrides]]
# Don't require test functions to include types
module = "tests.*"
allow_untyped_defs = true
disable_error_code = "attr-defined"
